+++ service/src/main/java/ru/practicum/ewm/event/service/EventService.java
@@ -40,5 +40,5 @@ public interface EventService {
 
     Event getEventById(Integer eventId);
 
-    Integer getEventListViews();
+    Long getEventListViews();
 }
